requirements.txt
discord.py
python-dotenv
google-genai
google-api-core
Pillow
pymongo